requirements/base.txt
Django==1.5.1
South==0.7.6
boto==2.9.2
bpython==0.12
dj-database-url==0.2.1
django-braces==1.0.0
django-model-utils==1.3.1
django-pipeline==1.3.9
# django-registration hasn't released 0.9.0 yet, but the code is good. Using this until they put it on PyPI.
-e hg+https://bitbucket.org/ubernostrum/django-registration@b3c41b3c7f50523b4ecb8267d6d27a918adc2e1f#egg=django_registration-dev
django-storages==1.1.8
django-taggit==0.10a1
logutils==0.3.3
psycopg2==2.5
